Implement eq.
One interesting problem vs. hand writing the assembly is that you need
to automatically generate unique labels. I've gone with a UUID, so you
end up with:

@EQ.<UUID>
(EQ.<UUID>)

Where <UUID> is a v4 UUID.

I see in Pong.asm of project 6 that a counter is used.

@EQ.<counter>
(EQ.<counter>)

That would be achievable in toAsm by passing in a counter in the 1st of
the curried arguments and incrementing. I suspect implementing lt and gt
will then require their own counters, so a probably a Set of counters
will be needed, assuming you want to count separately for each label,
which would be intuitive to a reader of asm, although not vital to make
it work.